Create keywordTrie.js
Extension of `TrieOfArrayUnique` class, right before the finish line. Currently edited keyword is yet to appear among `rxSplit` keyword results per each `updateSearchTerms` execution. The latter method is triggered on each keystroke in search of matching trie nodes.

Then, likely at `onchange` form input event or similar, cumulative results per `inputKeywords => term` need to be fetched from API which queries the underlying `TrieOfArrayUnique.#trie`; each unique terms, found in correlating with given word inputs, as they match in traversing tries (word-deep).

During trie recursion, while typing in input forms, as based on site features, new word nodes are prepended as to the declared sorting function in place, and are placed in a `new Array` per each term -- sorting of the `new index node` and placing `new words` within `TERMS<Array>` as they fall in their place (that is at place of using `splice` method, when the exact keyword is unique and doesn't appear on next). There is an execution breaking error you will find and is probably the first to be addressed, when attention is given to this kind of trivial challenge (and statistical methods are applied, other than mere frequency counting).

is non-trivial and essential This is hereby achieved with use of `sortAdd` method, querying for each `term`'s frequency in source data.

The method at instance, `parse`, is used to import datasets, then called collections, made of split string values with pointers into collections, named `this.TERMS`. Resulting from precalculated word frequencies from `parse` and `updateSearchTerms` methods, and narrowed down with each keystroke, appending as to the sorting function within. However, unique recombinations of `TERMS` are not currently stored and are to be recalculated and sorted, using `sortAdd` on each new path, and leafWordNode at `this.#terminator` symbol.

Besides, the `exportTrie` method may be used to communicate over network the state that results from parsing (`parse` in mention).

I think this one is promising because it's quite there. Well, perhaps the `rxSplit` method is oversimplistic, but for cases where URLs are unlikely to appear it is design to completely solve the problem of querying and caching.

Oh, yes, maybe there's one more thing in there left unused, unsolved.